Adicionado classe dao Veiculo e o arquivo da classe dao de Servico

+++ src/oficinamecanica/model/Servico.java
@@ -2,7 +2,7 @@ package oficinamecanica.model;
 public class Servico {
    private int id;
    private String descricao;
-   private float preco;
+   private double preco;
    
    //Construtores
     public Servico(){}
@@ -31,7 +31,7 @@ public class Servico {
         this.descricao = descricao;
     }
 
-    public float getPreco() {
+    public double getPreco() {
         return preco;
     }
